[BF] fix color float

+++ pertyG-1.0/Base/Graphic/Color.cpp
@@ -4,16 +4,16 @@ namespace pertyG
 {
     // Constructors
     Color::Color(Property r, Property g, Property b, Property a)
-        : red(r < 0.0f ? 0.0f : (r > 1.0f ? 1.0f : r)),
-        green(g < 0.0f ? 0.0f : (g > 1.0f ? 1.0f : g)),
-        blue(b < 0.0f ? 0.0f : (b > 1.0f ? 1.0f : b)),
-        alpha(a < 0.0f ? 0.0f : (a > 1.0f ? 1.0f : a)) {}
+        : red((double)r < 0.0 ? 0.0 : ((double)r > 1.0 ? 1.0 : r)),
+        green((double)g < 0.0 ? 0.0 : ((double)g > 1.0 ? 1.0 : g)),
+        blue((double)b < 0.0 ? 0.0 : ((double)b > 1.0 ? 1.0 : b)),
+        alpha((double)a < 0.0 ? 0.0 : ((double)a > 1.0 ? 1.0 : a)) {}
 
     Color::Color(int r, int g, int b, int a)
-        : red((r < 0 ? 0.0f : (r > 255 ? 1.0f : static_cast<double>(r) / 255.0f))),
-        green((g < 0 ? 0.0f : (g > 255 ? 1.0f : static_cast<double>(g) / 255.0f))),
-        blue((b < 0 ? 0.0f : (b > 255 ? 1.0f : static_cast<double>(b) / 255.0f))),
-        alpha((a < 0 ? 0.0f : (a > 255 ? 1.0f : static_cast<double>(a) / 255.0f))) {}
+        : red((r < 0 ? 0.0 : (r > 255 ? 1.0 : static_cast<double>(r) / 255.0))),
+        green((g < 0 ? 0.0 : (g > 255 ? 1.0 : static_cast<double>(g) / 255.0))),
+        blue((b < 0 ? 0.0 : (b > 255 ? 1.0 : static_cast<double>(b) / 255.0))),
+        alpha((a < 0 ? 0.0 : (a > 255 ? 1.0 : static_cast<double>(a) / 255.0))) {}
 
     // Getter methods
     Property & Color::getRed()  { return red; }
@@ -22,10 +22,10 @@ namespace pertyG
     Property & Color::getAlpha()  { return alpha; }
 
     // Setter methods
-    void Color::setRed(double r) { red = (r < 0.0f ? 0.0f : (r > 1.0f ? 1.0f : r)); }
-    void Color::setGreen(double g) { green = (g < 0.0f ? 0.0f : (g > 1.0f ? 1.0f : g)); }
-    void Color::setBlue(double b) { blue = (b < 0.0f ? 0.0f : (b > 1.0f ? 1.0f : b)); }
-    void Color::setAlpha(double a) { alpha = (a < 0.0f ? 0.0f : (a > 1.0f ? 1.0f : a)); }
+    void Color::setRed(double r) { red = (r < 0.0 ? 0.0 : (r > 1.0 ? 1.0 : r)); }
+    void Color::setGreen(double g) { green = (g < 0.0 ? 0.0 : (g > 1.0 ? 1.0 : g)); }
+    void Color::setBlue(double b) { blue = (b < 0.0 ? 0.0 : (b > 1.0 ? 1.0 : b)); }
+    void Color::setAlpha(double a) { alpha = (a < 0.0 ? 0.0 : (a > 1.0 ? 1.0 : a)); }
 
     // Method to get the brightness of the color
     double Color::getBrightness()
@@ -37,20 +37,20 @@ namespace pertyG
     std::string Color::toHex() 
     {
         std::stringstream ss;
-        ss << std::hex << std::setw(2) << std::setfill('0') << static_cast<int>(red * 255.0f);
-        ss << std::hex << std::setw(2) << std::setfill('0') << static_cast<int>(green * 255.0f);
-        ss << std::hex << std::setw(2) << std::setfill('0') << static_cast<int>(blue * 255.0f);
-        ss << std::hex << std::setw(2) << std::setfill('0') << static_cast<int>(alpha * 255.0f);
+        ss << std::hex << std::setw(2) << std::setfill('0') << static_cast<int>(red * 255.0);
+        ss << std::hex << std::setw(2) << std::setfill('0') << static_cast<int>(green * 255.0);
+        ss << std::hex << std::setw(2) << std::setfill('0') << static_cast<int>(blue * 255.0);
+        ss << std::hex << std::setw(2) << std::setfill('0') << static_cast<int>(alpha * 255.0);
         return ss.str();
     }
 
     // Method to blend two colors
     Color Color::blend( Color& other) 
     {
-        float blendedRed = (red + other.red) / 2.0f;
-        float blendedGreen = (green + other.green) / 2.0f;
-        float blendedBlue = (blue + other.blue) / 2.0f;
-        float blendedAlpha = (alpha + other.alpha) / 2.0f;
+        float blendedRed = (red + other.red) / 2.0;
+        float blendedGreen = (green + other.green) / 2.0;
+        float blendedBlue = (blue + other.blue) / 2.0;
+        float blendedAlpha = (alpha + other.alpha) / 2.0;
         return Color(blendedRed, blendedGreen, blendedBlue, blendedAlpha);
     }
